Исправил замечания по codestyle
