+++ app/src/main/java/ru/practicum/android/diploma/util/Debounce.kt
@@ -5,10 +5,12 @@ import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
-fun <T> debounce(delayMillis: Long,
-                 coroutineScope: CoroutineScope,
-                 useLastParam: Boolean,
-                 action: (T) -> Unit): (T) -> Unit {
+fun <T> debounce(
+    delayMillis: Long,
+    coroutineScope: CoroutineScope,
+    useLastParam: Boolean,
+    action: (T) -> Unit
+): (T) -> Unit {
     var debounceJob: Job? = null
     return { param: T ->
         if (useLastParam) {
