--- app/src/main/java/ru/practicum/android/diploma/util/Resourse.kt
@@ -1,6 +1,0 @@
-package ru.practicum.android.diploma.util
-
-sealed class Resource<T>(val data: T? = null, val message: String? = null) {
-    class Success<T>(data: T): Resource<T>(data)
-    class Error<T>(message: String, data: T? = null): Resource<T>(data, message)
-}
