+++ app/src/main/java/ru/practicum/android/diploma/util/network/NetworkClient.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.util.network
+
+interface NetworkClient {
+    suspend fun doRequest(dto: Any): Response
+}
