+++ app/src/main/java/ru/practicum/android/diploma/util/network/Response.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.util.network
+
+open class Response {
+    var resultCode = 0
+}
