--- app/src/main/java/ru/practicum/android/diploma/util/networkUtils/NetworkClient.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.util.networkUtils
-
-interface NetworkClient {
-   suspend fun doRequest(dto: Any): Response
-}
