--- app/src/main/java/ru/practicum/android/diploma/util/networkUtils/Response.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.util.networkUtils
-
-open class Response {
-    var resultCode = 0
-}
